Backport fix of CVE-2009-3300 for SP
Use internal functions instead of bumping the soname of the XMLTooling
and OpenSAML libraries as the original upstream fix did.  This also entails
hardwiring the newly introduced allowedSchemes configuration option to HTTP and
HTTPS only.  Unfortunately, the new symbols are not fully hidden, because various
components of this package use them from libshibsp.